Adicionando Swagger, logs com log4j e atualizando as regras de negócio com lambda

+++ src/main/java/br/com/shoppinglist/controller/ItemController.java
@@ -4,15 +4,18 @@ import br.com.shoppinglist.model.Item;
 import br.com.shoppinglist.model.ListaItem;
 import br.com.shoppinglist.model.TipoItem;
 import br.com.shoppinglist.service.ItemService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
-@RequestMapping("/shopping_list")
+@RequestMapping("/v1/item")
 @RestController
 public class ItemController {
     final ItemService itemService;
+    private static final Logger log = LoggerFactory.getLogger(ItemController.class);
 
     public ItemController(ItemService itemService) {
         this.itemService = itemService;
@@ -20,48 +23,64 @@ public class ItemController {
 
     // ============== ITEM ==============
 
-    @PostMapping("/item")
+    @PostMapping("/criar")
     public ResponseEntity<Object> saveItem(@RequestBody Item item) {
-        itemService.criarItem(item);
-        return ResponseEntity.status(HttpStatus.CREATED).body(item);
+        if(!itemService.existsByItemDesc(item)) {
+            return ResponseEntity.status(HttpStatus.CREATED).body(item);
+        } else {
+            log.error("Já existe um item com esta descrição.");
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Já existe um item com esta descrição");
+        }
     }
 
-    @GetMapping("/item")
+    @GetMapping("/buscar")
     public ResponseEntity<List<Item>> readAllItems() {
         return ResponseEntity.status(HttpStatus.OK).body(itemService.buscarItens());
     }
 
-    @GetMapping("/item/{id}")
-    public ResponseEntity<Item> readByIdItem(@PathVariable(value = "id") Integer id) {
-        return ResponseEntity.status(HttpStatus.OK).body(itemService.buscarItens().get(id));
+    @GetMapping("/buscar/item")
+    public ResponseEntity<Item> readByIdItem(@RequestParam(required = true)  Integer id) {
+        return ResponseEntity.status(HttpStatus.OK).body(itemService.readByIdItem(id));
     }
 
-    @DeleteMapping("/item/{id}")
-    public ResponseEntity<Void> deleteItem(@PathVariable Integer id) {
-        itemService.deleteItem(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    @DeleteMapping("/deletar/{id}")
+    public ResponseEntity<Object> deleteItem(@PathVariable Integer id) {
+        try {
+            boolean matchDelete = itemService.buscarItens().stream().anyMatch(e -> e.getTipoItemId().equals(id));
+            if (matchDelete) {
+                itemService.deleteItem(id);
+                log.warn("ITEM COM O ID {} FOI DELETADO.", id);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                log.error("Não foi encontrado um item com este ID");
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não foi encontrado um item com este ID");
+            }
+        } catch (Exception e) {
+            log.error(e.getMessage());
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+        }
     }
 
     @PutMapping("/item")
     public ResponseEntity<Item> updateItem(@RequestBody Item item) {
-        itemService.updateItem(item.getId());
+        itemService.updateItem(item);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     // ================ TIPO ITEM ================
 
-    @PostMapping("/tipo_item")
+    @PostMapping("/criar_tipo_item")
     public ResponseEntity<TipoItem> saveTipo(@RequestBody TipoItem tipoItem) {
         itemService.criarTipoItem(tipoItem);
         return ResponseEntity.status(HttpStatus.OK).body(tipoItem);
     }
 
-    @GetMapping("/tipo_item/{id}")
-    public ResponseEntity<TipoItem> readByIdTipo(@PathVariable (value = "id") Integer id) {
+    @GetMapping("/buscar_tipo_item/tipo_item")
+    public ResponseEntity<TipoItem> readByIdTipo(@RequestParam(required = true) Integer id) {
         return ResponseEntity.status(HttpStatus.OK).body(itemService.readByIdTipo(id));
     }
 
-    @GetMapping("tipo_item")
+    @GetMapping("/buscar_tipo_item")
     public ResponseEntity<List<TipoItem>> readAllTipos() {
         return ResponseEntity.status(HttpStatus.OK).body(itemService.buscarTipos());
     }
@@ -72,7 +91,7 @@ public class ItemController {
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @DeleteMapping("/tipo_item/{id}")
+    @DeleteMapping("/deletar_tipo_item/{id}")
     public ResponseEntity<Void> deleteTipo(@PathVariable Integer id) {
         itemService.deleteTipo(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
